+++ bin/dev-init/init.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+process.env.DEBUG = '0';
+require('../../module/boot/bin/init');
+require('./index').init();
