+++ module/boot/.gitignore
@@ -2,3 +2,5 @@ src/*.js
 src/*.d.ts
 index.d.ts
 index.js
+main.d.ts
+main.js
