--- module/boot/main.ts
@@ -1,5 +1,0 @@
-export * from './src/app-cache';
-export * from './src/cache';
-export * from './src/env';
-export * from './src/fs-util';
-export * from './src/register';
