+++ module/boot/src-ts/tsconfig.json
@@ -3,15 +3,16 @@
     "module": "commonjs",
     "target": "es2018",
     "strict": true,
+    "declaration": true,
     "strictPropertyInitialization": false,
-    "sourceMap": true,
+    "inlineSourceMap": true,
+    "outDir": "../src",
     "lib": [
       "es2015.proxy",
       "es2018"
     ],
     "preserveConstEnums": true,
     "experimentalDecorators": true,
-    "importHelpers": true,
-    "outDir": "build/"
+    "importHelpers": true
   }
 }
