+++ module/cli/src/launch.ts
@@ -1,3 +1,3 @@
 import { Execute } from './execute';
 
-export default Execute.run.bind(Execute);
+export const run = Execute.run.bind(Execute);
