+++ module/compiler/bin/compile-target.js
@@ -1,7 +1,6 @@
 require('@travetto/boot/bin/init');
-const { PhaseManager } = require(`@travetto/base`);
 
-PhaseManager.run().then(() => {
+require(`@travetto/base`).PhaseManager.run().then(() => {
   const { Compiler } = require('../src/compiler');
   return Compiler.compileAll();
 });
