+++ module/di/bin/find-apps.js
@@ -1,8 +1,7 @@
-const TRV_DI_ROOT = !process.env.TRV_FRAMEWORK_DEV ? '..' :
-  (process.cwd().includes('/module/di') ? process.cwd() :
-    `${process.cwd()}/node_modules/@travetto/di`);
+const cwd = process.cwd();
+require('@travetto/boot/bin/init');
 
-const { computeApps, handleFailure } = require(`${TRV_DI_ROOT}/bin/lib`);
-computeApps()
-  .then(resolved => require('fs').writeSync(1, `${JSON.stringify(resolved)}\n`))
-  .catch(err => handleFailure(err, 1));
+const TRV_DI_ROOT = !process.env.TRV_FRAMEWORK_DEV ? '..' :
+  (cwd.includes('/module/di') ? cwd : `${cwd}/node_modules/@travetto/di`);
+
+require(`${TRV_DI_ROOT}/bin/lib`).findApps();
