+++ module/schema/e2e/src/index.js
@@ -1,11 +1,10 @@
 require('@travetto/boot/bin/init');
-require('@travetto/base').PhaseManager
-  .run().then(() => {
-    require('./watch');
-    require('../src').SchemaRegistry.onFieldChange((e) => {
-      console.log('Field', e);
-    });
-    require('../src').SchemaRegistry.onSchemaChange((e) => {
-      console.log('Schema', e);
-    });
+require('@travetto/base').PhaseManager.run().then(() => {
+  require('./watch');
+  require('../src').SchemaRegistry.onFieldChange((e) => {
+    console.log('Field', e);
   });
+  require('../src').SchemaRegistry.onSchemaChange((e) => {
+    console.log('Schema', e);
+  });
+});
